Try to read hostname from env if not in updatedBuildData
If a job has no child stages (eg: in a freestyle job), updatedBuildData
will never actually be updated, so even if we are in a worker node
everything we read from it will be the main node's info. When this
happens, try to read the hostname from the DD_CI_HOSTNAME env var directly.